requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
